requirements/development.txt
Django==1.6.1
Pillow==2.3.0
South==0.8.4
WebOb==1.3.1
argparse==1.2.1
bobo==1.0.0
boto==2.23.0
django-storages==1.1.8
django-widget-tweaks==1.3
dropbox==2.0.0
path.py==5.0
pytz==2013.9
urllib3==1.7.1
wsgiref==0.1.2